requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
email-validator>=2.2.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
redis>=5.0.0
celery>=5.4.0
pandas>=2.2.0
openpyxl>=3.1.0
